Пользователи в рассылке прайсов в админке

+++ admin/templates/subscribePrices.php
@@ -1,4 +1,9 @@
 <?php
+
+/** @global Database $db  */
+
+use core\Database;
+
 $page_title = "Рассылка прайсов";
 $status = "<a href='/'>Главная</a> > Администрирование > ";
 require_once('pagination.php');
@@ -13,7 +18,12 @@ switch($_GET['act']){
 		header("Location: {$_SERVER['HTTP_REFERER']}");
 		break;
 	case 'delete':
-		$db->delete('subscribe_prices', "`email` = '{$_GET['email']}'");
+        if ($_GET['user_id']){
+            \core\User::update($_GET['user_id'], ['is_subscribe' => 0]);
+        }
+        else{
+            $db->delete('subscribe_prices', "`email` = '{$_GET['email']}'");
+        }
 		message('Успешно удалено');
 		header("Location: {$_SERVER['HTTP_REFERER']}");
 		break;
@@ -28,20 +38,46 @@ switch($_GET['act']){
 		$params['all'] = $db->getCount('subscribe_prices', $where);
 		$params['perPage'] = 30;
 		$linkLimit = 10;
-		$params['page'] = $_GET['page'] ? $_GET['page'] : 1;
-		$params['chank'] = getChank($params['all'], $params['perPage'], $linkLimit, $params['page']);
-		$start = $params['chank'][$params['page']] ? $params['chank'][$params['page']] : 0;
+		$params['page'] = $_GET['page'] ?: 1;
+		$params['chunk'] = getChank($params['all'], $params['perPage'], $linkLimit, $params['page']);
+		$start = $params['chunk'][$params['page']] ?: 0;
 		if ($where) $where = "WHERE $where";
 		$res_common_list = $db->query("
-			SELECT
-				*
-			FROM
-				#subscribe_prices
-			$where
-			ORDER BY
-				created DESC
-			LIMIT
-				$start, {$params['perPage']}
+			SELECT 
+			    '' as id,
+			    email,
+                title,
+                phone
+            FROM 
+                tahos_subscribe_prices
+            
+            UNION
+            
+            SELECT
+                u.id,                
+                email,
+                concat(
+                   IF(
+                       u.organization_name <> '',
+                       CONCAT_WS(' ', ot.title, u.organization_name),
+                       CONCAT_WS(' ', u.name_1, u.name_2, u.name_3)
+                    ),
+                    IF(
+                        @date := (select created from tahos_orders where user_id = u.id order by created desc limit 1),
+                        concat(
+                            ' (заказ <b>',
+                            @date,
+                            '</b>)'
+                        ),
+                        ''
+                    )
+                ) as title,
+                phone
+            FROM tahos_users u
+            LEFT JOIN tahos_organizations_types ot ON ot.id = u.organization_type
+            WHERE u.is_subscribe = 1
+            order by email
+            LIMIT $start, {$params['perPage']}
 		", '');
 		commonList($res_common_list, $params);
 }
@@ -63,12 +99,12 @@ function commonList($res_common_list, $params){?>
 		</tr>
 		<?if ($res_common_list->num_rows){
 			foreach($res_common_list as $item){?>
-				<tr class="edit">
+				<tr class="edit <?=$item['id'] ? 'user_id' : ''?>">
 					<td label="Email"><?=$item['email']?></td>
 					<td label="Название"><?=$item['title']?></td>
 					<td label="Телефон"><?=$item['phone']?></td>
 					<td label="">
-						<a tooltip="Удалить" class="delete" href="?view=subscribePrices&act=delete&email=<?=$item['email']?>">
+						<a tooltip="Удалить" class="delete" href="?view=subscribePrices&act=delete&email=<?=$item['email']?>&user_id=<?=$item['id']?>">
 							<span class="icon-cancel-circle1"></span>
 						</a>
 						<a target="_blank" email="<?=$item['email']?>" tooltip="Отправить вручную" class="subscribeHandy" href="#">
@@ -84,5 +120,5 @@ function commonList($res_common_list, $params){?>
 			</tr>
 		<?}?>
 	</table>
-	<?pagination($params['chank'], $params['page'], ceil($params['all'] / $params['perPage']), "?view=subscribePrices&search={$params['search']}&page=");?>
+	<?pagination($params['chunk'], $params['page'], ceil($params['all'] / $params['perPage']), "?view=subscribePrices&search={$params['search']}&page=");?>
 <?}?>
